requirements.txt
pandas
numpy
google-generativeai
python-dotenv
langchain==0.1.20 
langchain-core==0.1.52
langchain-community==0.0.38
langchain-google-genai==1.0.3
langchain-openai
